pyproject.toml
[project]
name = "ibkr-intraday-skeleton"
version = "0.1.0"
description = "IBKR intraday/micro-swing bot skeleton (no strategy)"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12",
  "PyYAML>=6.0",
  "pydantic>=2.5",
  "python-json-logger>=2.0",
  "ib-insync>=0.9.86",
  "tzdata>=2024.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
